Update TrinucleotideMatrix.R
Unless I misunderstood something or this is dealt with later in the code, believe there was a major bug in this function. The immediate 5' and 3' base of the mutated nucleotide must be the reverse-complemented for any purine. In the prior code, A[G>T]C was being converted to A[C>A]G etc. The correct conversion would be G[C>A]T, for this example. 
Rationale:
duplex for triplet
5'-AGC-3'
3'-TCG-5'

C is flanked by G(upstream, 5') and T (downstream, 3' side), hence GCT or G[C>A]T. Please feel free to reject this pull request if I'm mistaken and this is already handled properly elsewhere in the code.