Implement showMutationRate argument in lollipopPlot
`lollipopPlot` has a `showMutationRate` argument, which is declared not never explained nor implemented in any way. I perceive this argument to control whether the somatic mutation rate is displayed as part of the title, and implemented it accordingly.

This commit sets the plot title, saved as the variable `cbioSubtitle`, to be `geneID` as default; somatic mutation rate is appended to the title if `showMutationRate=TRUE`.