Update forestPlot.R
1 The lastest version maftools(v0.9.30) do not support color=c() paramter in forestPlot function.
re-Add color parameter which controls the color of the line depending on the location of the line (current x value is 1, will be 0 if using log(OR))
#' color=c('maroon','royalblue'): vector of colors or named vector of colors for each Variant_Classification
makes it compatible with older version of maftools.
2 change the annotation columns order of the plot, makes it concord with the group name order aprears on the left graphic.
eg. former will be group2/group1 in left graph and group1, group2, OR, P-value in right annotation,
after the change, the apprearance will be like
group2/group1 in left graph and group2, group1, OR, P-value in right annotation
formatted the code and add more annotations makes it easy to understand next time reading.